[CHORE]: Move options, deprecation utilities to their own modules in chainladder._config.
